Refactor: Extract NamedBuilder interface from EntityBuilder interface
